test(benchmarks): isolate runtime fixtures
