+++ benchmarks/benchmark_test.go
@@ -1,4 +1,4 @@
-package benchmarks
+package benchmarks_test
 
 import (
 	"context"
