+++ benchmarks/compare_test.go
@@ -1,6 +1,6 @@
 //go:build compare
 
-package benchmarks
+package benchmarks_test
 
 import (
 	"context"
